Clean-up section headers
The section header for the 4-stage cores had a typo (Embbbeded --> Embedded).  While I was add it I took the liberty of adding punctuation so that all headers are of the form, `XXXXX Class, N-Stage Core`.   The typo is a "must fix", but the altered heading format is a matter of personal taste which I am not wed to.